pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-podcast"
version = "0.1.0"
description = "Convert academic papers from Zotero into podcast-style audio reviews"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "python-dotenv>=1.0",
    "pymupdf>=1.24",
    "anthropic>=0.40",
    "elevenlabs>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
paper-podcast = "paper_podcast.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
